# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VitalMatch API - Connecting hospitals with blood donors in Kenya
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register/donor:
    post:
      operationId: registerDonor
      tags: [auth]
      summary: Register a new donor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDonorBody"
      responses:
        "201":
          description: Donor registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register/hospital:
    post:
      operationId: registerHospital
      tags: [auth]
      summary: Register a new hospital
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterHospitalBody"
      responses:
        "201":
          description: Hospital registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /appeals:
    post:
      operationId: createAppeal
      tags: [appeals]
      summary: Create a new emergency blood appeal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppealBody"
      responses:
        "201":
          description: Appeal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appeal"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /appeals/{id}:
    get:
      operationId: getAppeal
      tags: [appeals]
      summary: Get appeal details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appeal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppealWithHospital"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAppealStatus
      tags: [appeals]
      summary: Update appeal status (hospital only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppealStatusBody"
      responses:
        "200":
          description: Updated appeal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appeal"
  /appeals/{id}/respond:
    post:
      operationId: respondToAppeal
      tags: [appeals]
      summary: Accept or decline an appeal (donor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondToAppealBody"
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppealResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [donor, hospital]
      required:
        - email
        - password
        - role
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - name
        - createdAt
    RegisterDonorBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        phone:
          type: string
        dateOfBirth:
          type: string
      required:
        - name
        - email
        - password
        - bloodType
        - phone
        - dateOfBirth
    RegisterHospitalBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        address:
          type: string
        county:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
        - name
        - email
        - password
        - phone
        - address
        - county
        - latitude
        - longitude
...
    Appeal:
      type: object
      properties:
        id:
          type: integer
        hospitalId:
          type: integer
        bloodType:
          type: string
        quantity:
          type: integer
        emergencyLevel:
          type: string
          enum: [critical, urgent, standard]
        status:
          type: string
          enum: [active, fulfilled, cancelled]
        description:
          type: ["string", "null"]
        acceptedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - hospitalId
        - bloodType
        - quantity
        - emergencyLevel
        - status
        - description
        - acceptedCount
        - createdAt
        - expiresAt
    AppealWithHospital:
      type: object
      properties:
        id:
          type: integer
        hospitalId:
          type: integer
        hospitalName:
          type: string
        hospitalAddress:
          type: string
        hospitalLatitude:
          type: number
        hospitalLongitude:
          type: number
        bloodType:
          type: string
        quantity:
          type: integer
        emergencyLevel:
          type: string
          enum: [critical, urgent, standard]
        status:
          type: string
          enum: [active, fulfilled, cancelled]
        description:
          type: ["string", "null"]
        acceptedCount:
          type: integer
        distanceKm:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - hospitalId
        - hospitalName
        - hospitalAddress
        - hospitalLatitude
        - hospitalLongitude
        - bloodType
        - quantity
        - emergencyLevel
        - status
        - description
        - acceptedCount
        - distanceKm
        - createdAt
        - expiresAt
    CreateAppealBody:
      type: object
      properties:
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        quantity:
          type: integer
        emergencyLevel:
          type: string
          enum: [critical, urgent, standard]
        description:
          type: string
      required:
        - bloodType
        - quantity
        - emergencyLevel
    UpdateAppealStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, fulfilled, cancelled]
      required:
        - status
    RespondToAppealBody:
      type: object
      properties:
        action:
          type: string
          enum: [accept, decline]
      required:
        - action
    AppealResponse:
      type: object
      properties:
        id:
          type: integer
        appealId:
          type: integer
        donorId:
          type: integer
        action:
          type: string
        token:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - appealId
        - donorId
        - action
        - token
        - createdAt
...