# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VitalMatch API - Connecting hospitals with blood donors in Kenya
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  role: string;
  name: string;
  createdAt: Date;
}